"just" LTO the kernel hey
and the image is 1650K ... that would fit, except
that it likely is totally nonfunctional right now.

also try to have openwrt not muck up my kernel config ...
that didnt work, so more on that later.

seriously, need a full config file override i guess,
have had it with my config being twiddled with by
"the system".